requirements.txt

django
django-jazzmin
django-ckeditor
Pillow
